+++ source/Programs/Compositor/HDR/rays.frag
@@ -14,7 +14,7 @@
 SAMPLER2D(FBO, 0);
 uniform mediump float LightCount;
 uniform mediump vec4 LightPositionViewSpace[MAX_LIGHTS];
-uniform mediump int RayCount;
+uniform int RayCount;
 uniform mediump float Decay;
 uniform mediump float Density;
 uniform mediump float Weight;
@@ -22,7 +22,7 @@ uniform mediump float Exposure;
 
 
 //----------------------------------------------------------------------------------------------------------------------
-mediump vec3 GodRays(sampler2D tex, const mediump vec2 uv, const mediump vec2 lightPos, const mediump int counter, const mediump float density, const mediump float weight, const mediump float decay, const mediump float exposure)
+mediump vec3 GodRays(sampler2D tex, const mediump vec2 uv, const mediump vec2 lightPos, const int counter, const mediump float density, const mediump float weight, const mediump float decay, const mediump float exposure)
 {
     mediump vec3 color = vec3(0.0, 0.0, 0.0);
     mediump vec2 suv = uv;
